requirements.txt
astroid==2.0.4
Django==2.1.2
djangorestframework==3.8.2
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pep8==1.7.1
psycopg2-binary==2.7.5
pycodestyle==2.4.0
pytz==2018.5
six==1.11.0
wrapt==1.10.11
